backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
python-decouple>=3.6
djangorestframework-simplejwt>=5.2
channels>=4.0
celery>=5.3
Pillow>=10.0
django-extensions>=3.2
coverage>=7.2
black>=23.0
isort>=5.12
flake8>=6.0
pytest>=7.3
pytest-django>=4.5
factory-boy>=3.2

# Production dependencies (commented out for development)
# psycopg2-binary>=2.9
# channels-redis>=4.1  
# redis>=4.5
# WeasyPrint>=60.0